requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-Login==0.6.*
Flask-WTF==1.2.*
Flask-Migrate==4.0.*
Authlib==1.4.*
gunicorn
psycopg2-binary
python-dotenv
requests
beautifulsoup4
email-validator
google-genai
